bytecub-plugin/requirements.txt
argostranslate
azure-ai-translation-text==1.0.1
bitstring>=4.3.0
configargparse>=1.7
deepl
doclayout-yolo==0.0.3
# flask==2.2.3
# flask-cors==3.0.10
fonttools
freetype-py>=2.5.1
# gevent==22.10.2
gradio
gradio_pdf==0.0.21
huggingface-hub>=0.27.0
httpx[socks]>=0.27.0
modelscope
msgpack>=1.1.0
numpy>=2.0.2
ollama
onnx>=1.17.0
onnxruntime>=1.16.1
openai>=1.59.3
opencv-python-headless
opencv-python>=4.10.0.84
orjson>=3.10.14
peewee>=3.17.8
pikepdf
pymupdf>=1.25.1
python-dotenv
python-levenshtein>=0.27.1
pdfminer.six>=20240706
pydantic>=2.10.6
requests
rich>=13.9.4
scikit-image>=0.25.2
tenacity>=9.0.0
tencentcloud-sdk-python
tiktoken>=0.9.0
toml>=0.10.2
torch==2.5.1
torchvision
tqdm>=4.67.1
werkzeug==2.2.2
xsdata[cli,lxml,soap]>=24.12
xinference-client
rapidocr-onnxruntime>=1.3.7
requests==2.31.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.1,<0.28.0
httpx==0.27.2
sse-starlette==2.0.0
markdown==3.8
python-docx==1.1.2
beautifulsoup4==4.13.4
pylatexenc==2.10
matplotlib==3.10.1
pillow==11.2.1
pyzstd>=0.17.0